requirements.txt
# pip install -r requirements.txt
python-dotenv
python-dateutil
python-binance
openai
pandas
numpy
requests
streamlit
altair